refactor: entity paths.
