+++ dialogy/plugins/text/slot_filler/rule_slot_filler.py
@@ -6,7 +6,7 @@ Module provides access to a rule-based :ref:`slot filler<slot_filler>`.
 from typing import Any, List, Optional
 
 from dialogy.base import Guard, Input, Output, Plugin
-from dialogy.types.entity import BaseEntity
+from dialogy.types import BaseEntity
 from dialogy.types.intent import Intent
 from dialogy.types.slots import Rule
 from dialogy.utils.logger import logger
