+++ dialogy/types/__init__.py
@@ -25,17 +25,17 @@ We use :ref:`slots<slot>` within :ref:`intents<intent>` to associate relationshi
 So these :ref:`entities<base_entity>` would be filled within their respective :ref:`slots<slot>` already expected by the
 reservation :ref:`intents<intent>`.
 """
-from dialogy.types.entity import (
-    BaseEntity,
-    CreditCardNumberEntity,
-    CurrencyEntity,
-    DurationEntity,
-    KeywordEntity,
-    NumericalEntity,
-    PeopleEntity,
-    TimeEntity,
-    TimeIntervalEntity,
-)
+from dialogy.types.entity.amount_of_money import CurrencyEntity
+from dialogy.types.entity.base_entity import BaseEntity, entity_synthesis
+from dialogy.types.entity.credit_card_number import CreditCardNumberEntity
+from dialogy.types.entity.duration import DurationEntity
+from dialogy.types.entity.keyword import KeywordEntity
+from dialogy.types.entity.numerical import NumericalEntity
+from dialogy.types.entity.people import PeopleEntity
+from dialogy.types.entity.time import TimeEntity
+from dialogy.types.entity.time_interval import TimeIntervalEntity
+from dialogy.types.entity.deserialize import deserialize_duckling_entity
+
 from dialogy.types.intent import Intent
 from dialogy.types.plugin import PluginFn
 from dialogy.types.signal.signal import Signal
