+++ dialogy/types/entity/duration/__init__.py
@@ -9,7 +9,7 @@ from typing import Any, Dict, Optional
 import attr
 
 from dialogy import constants as const
-from dialogy.types.entity import BaseEntity
+from dialogy.types import BaseEntity
 
 
 @attr.s
