+++ dialogy/types/entity/numerical/__init__.py
@@ -12,7 +12,7 @@ from typing import Any, Dict, Optional
 import attr
 
 from dialogy import constants as const
-from dialogy.types.entity import BaseEntity
+from dialogy.types.entity.base_entity import BaseEntity
 
 
 @attr.s
